+++ crates/ra_assists/src/lib.rs
@@ -100,7 +100,7 @@ mod split_import;
 mod remove_dbg;
 pub mod auto_import;
 mod add_missing_impl_members;
-mod move_guard_to_arm_body;
+mod move_guard_to_arm_body_and_back;
 
 fn all_assists<DB: HirDatabase>() -> &'static [fn(AssistCtx<DB>) -> Option<Assist>] {
     &[
@@ -119,7 +119,7 @@ fn all_assists<DB: HirDatabase>() -> &'static [fn(AssistCtx<DB>) -> Option<Assis
         add_missing_impl_members::add_missing_impl_members,
         add_missing_impl_members::add_missing_default_members,
         inline_local_variable::inline_local_varialbe,
-        move_guard_to_arm_body::move_guard_to_arm_body,
+        move_guard_to_arm_body_and_back::move_guard_to_arm_body,
     ]
 }
