+++ crates/ra_assists/src/move_guard_to_arm_body_and_back.rs
